Update openapi.yaml to make sure ending_at is always serialized in request
Imagine you are updating a subscription and want to remove the `ends_at` date.

Right now this doesn't work correctly: Settings `ends_at` to null in Java code will omit `ends_at` from the request all together.
The cause is that generated client specifies `@JsonInclude(Include.USE_DEFAULTS)` on the field, which will by default omit any null values from the request.

We need to override this for this specific field. Setting it to required in the spec will change the annotation to `@JsonInclude(Include.ALWAYS)` , which will cause it be serialised in the request, even if it's value is null.