feat(wallet-transaction): Add a `name` to wallet transactions (#439)
## Context

When creating a wallet transaction, the name displayed on the invoice defaults to the wallet name. However, for organizations managing multiple types of top-ups, it would be beneficial to display specific naming on invoices that corresponds to the particular transaction being made.

## Description 

This is the documentation for changes introduced in https://github.com/getlago/lago-api/commit/eecb90fb60190504307c6d71337519e8fcf23a52. It adds a `name` attribute to the wallet transaction object and inputs. It also updates the description of the fees'a fields.