misc(charge): Make `charge_model` re-usable and add `charge_model` to subscriptions (#419)
## Context 

The `charge_model` is defined in multiple places and some of them are outdated. Also the subscription creation and update were missing this property in `plan_overrides`.

## Description

This adds a re-usable `src/schemas/ChargeModelEnum.yaml` file and make sure subscription creation and update includes the `charge_model` in `plan_overrides`.