feat(wallet-transaction): Add `transaction_name` to recurring transaction rules (#445)
## Context

When creating a wallet transaction, the name displayed on the invoice defaults to the wallet name. However, for organizations managing multiple types of top-ups, it would be beneficial to display specific naming on invoices that corresponds to the particular transaction being made.

This is related to https://github.com/getlago/lago-api/commit/38261baa273514e1b79b8c9be5263f4ef21f0eed.

## Description 

This change adds a `name` attribute to the wallet transaction recurring rule object and inputs.